Tests - Sanity - Improved Publish / Unpublish tests
*Added test cases for publsih / unpublish with block / filesystem and added logic to make sure the mount paths arer created after publish and deleted after unpublsih

+++ test/sanity/helpers/container_server_manager.py
@@ -200,8 +200,9 @@ class ContainerServerManager(object):
 		device_file = os.path.join(dev_nvmesh, volume_id)
 
 		self._make_sure_dir_exists(dev_nvmesh)
-		with open(device_file, 'w') as fp:
-			fp.write('')
+		if not os.path.isfile(device_file):
+			with open(device_file, 'w') as fp:
+				fp.write('')
 
 	def remove_nvmesh_device(self, volume_id):
 		dev_nvmesh = os.path.join(self.env_dir, 'dev/nvmesh')
